requirements.txt
argon2-cffi==20.1.0
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
bleach==4.0.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
colorama==0.4.4
cssselect2==0.4.1
cycler==0.10.0
debugpy==1.4.1
decorator==4.4.2
defusedxml==0.7.1
entrypoints==0.3
idna==3.2
imageio==2.9.0
imageio-ffmpeg==0.4.5
importlib-metadata==4.8.1
ipykernel==6.2.0
ipython==7.27.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.0.1
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.2
lxml==4.6.3
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.2
mistune==0.8.4
mkl-fft==1.3.0
mkl-random==1.2.2
mkl-service==2.4.0
moviepy==1.0.3
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.3
numpy==1.20.3
olefile==0.46
packaging==21.0
pandas==1.3.3
pandocfilters==1.4.3
parso==0.8.2
pickleshare==0.7.5
Pillow==8.3.2
pip==21.2.4
proglog==0.1.9
prometheus-client==0.11.0
prompt-toolkit==3.0.17
pycparser==2.20
Pygments==2.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.1
pywin32==228
pywinpty==0.5.7
pyzmq==22.2.1
qtconsole==5.1.0
QtPy==1.10.0
reportlab==3.6.1
requests==2.26.0
Send2Trash==1.5.0
setuptools==58.0.4
sip==4.19.13
six==1.16.0
svg.path==4.1
svglib==1.1.0
terminado==0.9.4
testpath==0.5.0
tinycss2==1.1.0
tornado==6.1
tqdm==4.62.3
traitlets==5.0.5
urllib3==1.26.7
wcwidth==0.2.5
webencodings==0.5.1
wheel==0.37.0
widgetsnbextension==3.5.1
wincertstore==0.2
zipp==3.5.0
